make StatisticsController an interface due to a all-open issue (private setters for open properties are prohibited, e.g. rank: Int)
